Simplify all the components module
