+++ src/pyfx/view/components/__init__.py
@@ -1,4 +1,6 @@
 """UI components (such as `query bar`) used in Pyfx's TUI."""
-from .autocomplete_popup import AutoCompletePopUp
-from .query_bar import QueryBar
-from .json_browser import JSONBrowser
+from pyfx.view.components.autocomplete_popup import AutoCompletePopUp
+from pyfx.view.components.help_popup import HelpPopUp
+from pyfx.view.components.json_browser import JSONBrowser
+from pyfx.view.components.query_bar import QueryBar
+from pyfx.view.components.warning_bar import WarningBar
