--- src/pyfx/view/components/autocomplete_popup/__init__.py
@@ -1,1 +1,0 @@
-from .autocomplete_popup import AutoCompletePopUp
