+++ src/pyfx/view/components/help_popup.py
@@ -1,10 +1,12 @@
+from dataclasses import dataclass
 from enum import Enum
 
 import urwid
 from overrides import overrides
 
-from ...common import SelectableText
-from ...keymapper import KeyDefinition
+from pyfx.view.common import SelectableText
+from pyfx.view.keymapper import AbstractComponentKeyMapper
+from pyfx.view.keymapper import KeyDefinition
 
 
 class HelpPopUpKeys(KeyDefinition, Enum):
@@ -15,6 +17,41 @@ class HelpPopUpKeys(KeyDefinition, Enum):
     @classmethod
     def list(cls):
         return list(map(lambda k: k.key, cls))
+
+
+@dataclass(frozen=True)
+class HelpPopUpKeyMapper(AbstractComponentKeyMapper):
+    cursor_up: str = "up"
+    cursor_down: str = "down"
+    exit: str = "esc"
+
+    @property
+    @overrides
+    def mapped_key(self):
+        return {
+            self.cursor_up: HelpPopUpKeys.CURSOR_UP,
+            self.cursor_down: HelpPopUpKeys.CURSOR_DOWN,
+            self.exit: HelpPopUpKeys.EXIT,
+        }
+
+    @property
+    @overrides
+    def short_help(self):
+        return [f"UP: {self.cursor_up}",
+                f"DOWN: {self.cursor_down}",
+                f"CLOSE: {self.exit}"]
+
+    @property
+    @overrides
+    def detailed_help(self):
+        keys = [
+            self.cursor_up, self.cursor_down, self.exit
+        ]
+        descriptions = {key: self.mapped_key[key].description for key in keys}
+        return {
+            "section": "Help Page",
+            "description": descriptions
+        }
 
 
 class HelpPopUp(urwid.WidgetWrap):
