--- src/pyfx/view/components/json_browser/__init__.py
@@ -1,1 +1,0 @@
-from .json_browser import JSONBrowser
