--- src/pyfx/view/components/query_bar/__init__.py
@@ -1,1 +1,0 @@
-from .query_bar import QueryBar
